Clarify units
On the internet the usage of b and B for bits or bytes is inconsistent.

Lowercase b seems to be used for either, and upperace B is more often
used to mean bytes but usege for bits also exists.

I am not aware of any standard that would mandate one or other use.

Consequently, spelling out the whole word in the output or adding
explanation in documentation is the only way to make sure the output is
interpreted correctly.

To save space on small screens only alter documentation and the bytes
label without prefixes/suffixes.

While at it change uppercase K to lowercase k to correctly spell the
kilo- prefix.

Fixes: #261

Signed-off-by: Michal Suchanek <msuchanek@suse.de>