determineVersion.sh: prevent traversing into outer git repo for git-describe call
When building nethogs inside another git repo (as is i.e. done as part
of a buildroot build) the version is picked up from the outer git repo,
which is confusing as well as introducing reproducibility problems when
rebuilding the same nethogs version produces different results if the
outer repository changes.